Refactor tradeup builder hook into modular helpers
