+++ client/src/modules/tradeups/utils/format.ts
@@ -1,0 +1,5 @@
+export const formatNumber = (value: number, digits = 2) =>
+  Number.isFinite(value) ? value.toFixed(digits) : "—";
+
+export const formatPercent = (value: number) =>
+  Number.isFinite(value) ? `${(value * 100).toFixed(2)}%` : "—";
